Remove bootcamps, csail, mitpe, see
